Enhance component service and update AddCommand
Code has been updated to allow multiple selection of components to be added when none are provided initially in the AddCommand. This eliminates the need to specify component names one by one. Additionally, a minor tweak was made to the ComponentService to fix a potential issue where only the first matching component would be returned.